gdl_web/requirements.txt
asgiref==3.11.1
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
coverage==7.13.4
cryptography==46.0.5
cssselect2==0.9.0
Django==6.0.3
dj-database-url==2.1.0
django-allauth==65.15.0
django-htmx==1.21.0
factory_boy==3.3.3
Faker==40.8.0
fonttools==4.62.0
gunicorn==23.0.0
idna==3.11
iniconfig==2.3.0
oauthlib==3.3.1
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
psycopg==3.2.13
psycopg-binary==3.2.13
pycparser==3.0
pydyf==0.12.1
Pygments==2.19.2
PyJWT==2.11.0
pyphen==0.17.2
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.1.0
urllib3==2.6.3
weasyprint==63.1
webencodings==0.5.1
zopfli==0.4.1
whitenoise==6.6.0
